dataPrediction-kafkaModifying/requirements.txt
amqp==5.1.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
blinker==1.5
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
findspark==2.0.1
Flask==2.2.2
fonttools==4.37.3
h5py==3.7.0
idna==3.3
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kafka-python==2.0.2
kit==0.2.15
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pycparser==2.21
Pygments==2.13.0
pymongo==4.3.2
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
scipy==1.9.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.1
terminado==0.15.0
tinycss2==1.1.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
